requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
streamlit==1.31.0
plotly==5.18.0
altair==5.2.0
python-dateutil==2.8.2
tqdm==4.66.1
